requirements.txt
opencv-python==3.4.0.14
numpy
tensorflow-gpu
matplotlib